Minor fixes to LeNet
