+++ src/LeNet_5/LeNet5.py
@@ -1,0 +1,30 @@
+from torch.nn import Module
+from torch import nn
+from torchinfo import summary
+from torch.nn import functional as F
+
+class LeNet5(Module):
+    def __init__(self):
+        super(LeNet5, self).__init__()
+        self.conv1 = nn.Conv2d(1, 6, 5, 1, padding=2)
+        self.conv2 = nn.Conv2d(6, 16, 5)
+        self.fc1 = nn.Linear(16*5*5, 120)
+        self.fc2 = nn.Linear(120, 84)
+        self.fc3 = nn.Linear(84, 10)
+
+    def forward(self, x):
+        y = F.relu(self.conv1(x))
+        y = F.max_pool2d(y, 2)
+        y = F.relu(self.conv2(y))
+        y = F.max_pool2d(y, 2)
+        y = y.view(y.size(0), -1)
+        y = F.relu(self.fc1(y))
+        y = F.relu(self.fc2(y))
+        y = self.fc3(y)
+        return y
+
+if __name__ == '__main__':
+    model = LeNet5()
+    # Summary of the model
+    print(summary(model=model, input_size=(1, 1, 28, 28), col_width=20,
+            col_names=['input_size', 'output_size', 'num_params', 'trainable'], row_settings=['var_names'], verbose=0))
