+++ src/LeNet_5/__init__.py
@@ -1,0 +1,6 @@
+
+from .LeNet5 import LeNet5
+from .train import Trainer
+from .test import Tester
+
+__all__ = ['LeNet5', 'Trainer', 'Tester']
